+++ napari/utils/_injection/_processors.py
@@ -1,61 +1,16 @@
 import sys
 from concurrent.futures import Future
 from contextlib import nullcontext, suppress
-from functools import partial, wraps
-from inspect import isgeneratorfunction, signature
+from functools import partial
 from typing import Any, Callable, Dict, Optional, Set, Type, TypeVar
 
-from typing_extensions import get_type_hints
-
-from .. import components, layers, types, viewer
-from ..layers._source import layer_source
-from ..utils.misc import ensure_list_of_layer_data_tuple
+from ... import layers, types, viewer
+from ...layers._source import layer_source
+from ...utils.misc import ensure_list_of_layer_data_tuple
+from ._accessors import _access_viewer
 
 T = TypeVar("T")
 _NULL = object()
-
-
-def _access_viewer() -> Optional[viewer.Viewer]:
-    return get_accessor(viewer.Viewer)()
-
-
-def _get_active_layer() -> Optional[layers.Layer]:
-    return v.layers.selection.active if (v := _access_viewer()) else None
-
-
-def _get_active_layer_list() -> Optional[components.LayerList]:
-    return v.layers if (v := _access_viewer()) else None
-
-
-# registry of Type -> "accessor function"
-# where each value is a function that is capable
-# of retrieving an instance of its corresponding key type.
-_ACCESSORS: Dict[Type, Callable[..., Optional[object]]] = {
-    layers.Layer: _get_active_layer,
-    viewer.Viewer: viewer.current_viewer,
-    components.LayerList: _get_active_layer_list,
-}
-
-
-def get_accessor(type_: Type[T]) -> Optional[Callable[..., Optional[T]]]:
-    """Return object accessor function given a type.
-
-    An object accessor is a function that returns an instance of a
-    particular object type. For example, given type `napari.Viewer`, we return
-    a function that can be called to get the current viewer.
-
-    This is a form of dependency injection, and, along with
-    `inject_napari_dependencies`, allows us to inject current napari objects
-    into functions based on type hints.
-    """
-    if type_ in _ACCESSORS:
-        return _ACCESSORS[type_]
-
-    if isinstance(type_, type):
-        for key, val in _ACCESSORS.items():
-            if issubclass(type_, key):
-                return val  # type: ignore [return-type]
-    return None
 
 
 def _add_layer_data_tuples_to_viewer(
@@ -205,7 +160,7 @@ def _add_future_data(
 
 
 # add default processors
-_PROCESSORS = {
+_PROCESSORS: Dict[Any, Callable[[Any], Any]] = {
     layers.Layer: _add_layer_to_viewer,
     types.LayerDataTuple: _add_layer_data_tuples_to_viewer,
 }
@@ -285,154 +240,3 @@ class set_processor:
                 del _PROCESSORS[key]
             else:
                 _PROCESSORS[key] = val
-
-
-class set_accessor:
-    """Set accessor(s) for given type(s).
-
-    "Accessors" are functions that can retrieve an instance of a given type.
-    For instance, `napari.viewer.current_viewer` is a function that can
-    retrieve an instance of `napari.Viewer`.
-
-    This is a class that behaves as a function or a context manager, that
-    allows one to set an accessor function for a given type.
-
-    Parameters
-    ----------
-    mapping : Dict[Type[T], Callable[..., Optional[T]]]
-        a map of type -> accessor function, where each value is a function
-        that is capable of retrieving an instance of the associated key/type.
-    clobber : bool, optional
-        Whether to override any existing accessor function, by default False.
-
-    Raises
-    ------
-    ValueError
-        if clobber is `False` and one of the keys in `mapping` is already
-        registered.
-    """
-
-    def __init__(
-        self, mapping: Dict[Type[T], Callable[..., Optional[T]]], clobber=False
-    ):
-        self._before = {}
-        for k in mapping:
-            if k in _ACCESSORS and not clobber:
-                raise ValueError(
-                    f"Class {k} already has an accessor and clobber is False"
-                )
-            self._before[k] = _ACCESSORS.get(k, _NULL)
-        _ACCESSORS.update(mapping)
-
-    def __enter__(self):
-        return None
-
-    def __exit__(self, *_):
-        for key, val in self._before.items():
-            if val is _NULL:
-                del _ACCESSORS[key]
-            else:
-                _ACCESSORS[key] = val
-
-
-def napari_type_hints(obj: Any) -> Dict[str, Any]:
-    """variant of get_type_hints with napari namespace awareness."""
-    import napari
-
-    return get_type_hints(
-        obj,
-        {
-            'napari': napari,
-            **viewer.__dict__,
-            **layers.__dict__,
-            **components.__dict__,
-        },
-    )
-
-
-def inject_napari_dependencies(func: Callable[..., T]) -> Callable[..., T]:
-    """Create callable that can access napari objects based on type hints.
-
-    This is form of dependency injection.  If a function includes a parameter
-    that has a recognized napari type (e.g. `Viewer`, or `Layer`), then this
-    function will return a new version of the input function that can be called
-    *without* that particular parameter.
-
-    Examples
-    --------
-
-    >>> def f(viewer: Viewer): ...
-    >>> inspect.signature(f)
-    <Signature (x: 'Viewer')>
-    >>> f2 = inject_napari_dependencies(f)
-    >>> inspect.signature(f2)
-    <Signature (x: typing.Optional[napari.Viewer] = None)>
-    # if f2 is called without x, the current_viewer will be provided for x
-
-
-    Parameters
-    ----------
-    func : Callable
-        A function with napari type hints.
-
-    Returns
-    -------
-    Callable
-        A function with napari dependencies injected
-    """
-    if not func.__code__.co_argcount and 'return' not in getattr(
-        func, '__annotations__', {}
-    ):
-        return func
-
-    sig = signature(func)
-    # get type hints for the object, with forward refs of napari hints resolved
-    hints = napari_type_hints(func)
-    # get accessor functions for each required parameter
-    required = {}
-    return_hint = None
-    for name, hint in hints.items():
-        if name == 'return':
-            return_hint = hint
-            continue
-        if sig.parameters[name].default is sig.empty:
-            required[name] = hint
-
-    @wraps(func)
-    def _exec(*args, **kwargs):
-        # when we call the function, we call the accessor functions to get
-        # the current napari objects
-        _kwargs = {}
-        for n, hint in required.items():
-            if accessor := get_accessor(hint):
-                _kwargs[n] = accessor()
-
-        # but we use bind_partial to allow the caller to still provide
-        # their own objects if desired.
-        # (i.e. the injected deps are only used if needed)
-        _kwargs.update(**sig.bind_partial(*args, **kwargs).arguments)
-        result = func(**_kwargs)
-        if return_hint and (processor := get_processor(return_hint)):
-            processor(result)
-        return result
-
-    out = _exec
-
-    # if it came in as a generatorfunction, it needs to go out as one.
-    if isgeneratorfunction(func):
-
-        @wraps(func)
-        def _gexec(*args, **kwargs):
-            yield from _exec(*args, **kwargs)
-
-        out = _gexec
-
-    # update the signature
-    p = [
-        p.replace(default=None, annotation=Optional[hints[p.name]])
-        if p.name in required
-        else p
-        for p in sig.parameters.values()
-    ]
-    out.__signature__ = sig.replace(parameters=p)
-    return out
